Mejoro estilos

+++ src/app/dashboard/partidos/editar/delete-btn.jsx
@@ -1,5 +1,7 @@
 "use client";
 
+import { TrashIcon } from "@/app/components/icons";
+import { Spinner } from "@/app/components/spinner";
 import { deleteMatch } from "@/app/lib/matches";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
@@ -10,13 +12,15 @@ export function DeleteBtn({ match_id }) {
 
   async function handleDelete(match_id) {
     setLoading(true);
-    const { error } = await deleteMatch(match_id);
+    if (confirm("Estás seguro de eliminar este partido?")) {
+      const { error } = await deleteMatch(match_id);
 
-    if (error) {
-      alert("Ocurrió un error. Intenta de nuevo más tarde...");
-    } else {
-      alert("Partido eliminado");
-      router.refresh();
+      if (error) {
+        alert("Ocurrió un error. Intenta de nuevo más tarde...");
+      } else {
+        alert("Partido eliminado");
+        router.refresh();
+      }
     }
 
     setLoading(false);
@@ -24,7 +28,7 @@ export function DeleteBtn({ match_id }) {
 
   return (
     <button onClick={async () => await handleDelete(match_id)} type="button">
-      {loading ? "Eliminando..." : "Eliminar"}
+      {loading ? <Spinner /> : <TrashIcon />}
     </button>
   );
 }
